+++ enagement/src/main/java/service/CSFC/CSFC_auth_service/model/dto/response/CouponResponse.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import service.CSFC.CSFC_auth_service.model.constants.DiscountType;
 import service.CSFC.CSFC_auth_service.model.entity.LoyaltyTier;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -19,9 +20,9 @@ public class CouponResponse {
     private Long promotionId;
     private String code;
     private DiscountType discountType;
-    private Double discountValue;
-    private Double minOrderValue;
-    private Double maxDiscount;
+    private BigDecimal discountValue;
+    private BigDecimal minOrderValue;
+    private BigDecimal maxDiscount;
     private Integer usageLimit;
     private Integer userLimit;
     private Integer usedCount;
